Consolidate Python and tool shell updaters
